Create `Heading` component
