+++ components/Culture.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+export default function Culture(props) {
+    return (
+        <View>
+            <Text>Culture</Text>
+        </View>
+    )
+}
