+++ components/Featured.js
@@ -1,0 +1,11 @@
+import React from "react"
+import { View, StyleSheet, Text, Image } from "react-native"
+
+export default function Featured(props) {
+
+    return (
+        <View>
+            <Text>Featured</Text>
+        </View>
+    )
+}
