+++ components/Heading.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+export default function Heading(props) {
+    return (
+        <View>
+            <Text>Heading</Text>
+        </View>
+    )
+}
